Refactor SAST passe results logic from FlowService into SastScannerService
