+++ src/main/java/com/checkmarx/flow/service/SastScannerService.java
@@ -164,6 +164,20 @@ public class SastScannerService {
         }
     }
 
+    public void cxParseResults(ScanRequest request, File file) throws ExitThrowable {
+        try {
+            ScanResults results = cxService.getReportContent(file, request.getFilters());
+            resultsService.processResults(request, results, scanDetails);
+            if(flowProperties.isBreakBuild() && results !=null && results.getXIssues()!=null && !results.getXIssues().isEmpty()){
+                log.error(ERROR_BREAK_MSG);
+                exit(10);
+            }
+        } catch (MachinaException | CheckmarxException e) {
+            log.error("Error occurred while processing results file", e);
+            exit(3);
+        }
+    }
+
     private String determineTeamAndOwnerID(ScanRequest request) throws CheckmarxException, MachinaException {
 
         String ownerId;
