+++ src/test/java/com/checkmarx/flow/cucumber/integration/azure/publishing/issueprocessing/PublishingSteps.java
@@ -12,6 +12,7 @@ import com.checkmarx.flow.dto.Issue;
 import com.checkmarx.flow.dto.ScanRequest;
 import com.checkmarx.flow.exception.ExitThrowable;
 import com.checkmarx.flow.service.FlowService;
+import com.checkmarx.flow.service.SastScannerService;
 import com.checkmarx.sdk.config.CxProperties;
 import io.cucumber.java.Before;
 import io.cucumber.java.en.And;
@@ -54,6 +55,9 @@ public class PublishingSteps extends PublishingStepsBase {
 
     @Autowired
     private AzureDevopsClient adoClient;
+
+    @Autowired
+    private SastScannerService sastScannerService;
 
     private String sastReportFilename;
 
@@ -142,7 +146,7 @@ public class PublishingSteps extends PublishingStepsBase {
         File sastReport = TestUtils.getFileFromResource(sastReportPath);
 
         try {
-            flowService.cxParseResults(request, sastReport);
+            sastScannerService.cxParseResults(request, sastReport);
         } catch (Throwable e) {
             if (expectingException) {
                 lastCxFlowException = e;
