+++ src/test/java/com/checkmarx/flow/cucumber/integration/github/GitHubCommonSteps.java
@@ -4,6 +4,7 @@ import com.checkmarx.flow.config.FlowProperties;
 import com.checkmarx.flow.dto.BugTracker;
 import com.checkmarx.flow.dto.ScanRequest;
 import com.checkmarx.flow.service.FlowService;
+import com.checkmarx.flow.service.SastScannerService;
 import com.checkmarx.flow.utils.github.GitHubTestUtilsImpl;
 import com.checkmarx.sdk.config.CxProperties;
 import com.checkmarx.sdk.dto.Filter;
@@ -33,6 +34,9 @@ public abstract class GitHubCommonSteps {
 
     @Autowired
     protected GitHubTestUtilsImpl gitHubTestUtils;
+
+    @Autowired
+    protected SastScannerService sastScannerService;
 
     protected File getFileFromResourcePath(String path) throws IOException {
         return new ClassPathResource(path).getFile();
