+++ src/test/java/com/checkmarx/flow/cucumber/integration/github/GitHubOpenIssuesBySeverityFiltersSteps.java
@@ -54,7 +54,7 @@ public class GitHubOpenIssuesBySeverityFiltersSteps extends GitHubCommonSteps {
     @When("publishing {string}")
     public void publishResults(String input) throws IOException, ExitThrowable {
         scanRequest = getBasicScanRequest();
-        flowService.cxParseResults(scanRequest, getFileFromResourcePath(INPUT_BASE_PATH + input));
+        sastScannerService.cxParseResults(scanRequest, getFileFromResourcePath(INPUT_BASE_PATH + input));
     }
 
     @Then("{int} new issues should be opened according filters severity")
