+++ src/test/java/com/checkmarx/flow/cucumber/integration/negative_tests/JiraUnreachableSteps.java
@@ -6,8 +6,8 @@ import com.checkmarx.flow.config.JiraProperties;
 import com.checkmarx.flow.dto.BugTracker;
 import com.checkmarx.flow.dto.ScanRequest;
 import com.checkmarx.flow.exception.JiraClientRunTimeException;
-import com.checkmarx.flow.service.FlowService;
 import com.checkmarx.flow.service.JiraService;
+import com.checkmarx.flow.service.SastScannerService;
 import com.checkmarx.sdk.config.Constants;
 import com.checkmarx.sdk.config.CxProperties;
 import io.cucumber.java.Before;
@@ -42,10 +42,10 @@ public class JiraUnreachableSteps {
     private CxProperties cxProperties;
 
     @Autowired
-    private FlowService flowService;
+    private JiraService jiraService;
 
     @Autowired
-    private JiraService jiraService;
+    private SastScannerService sastScannerService;
 
     private BugTracker bugTracker;
 
@@ -74,7 +74,7 @@ public class JiraUnreachableSteps {
     @Then("the call execution should throw a JiraClientRunTimeException since JIRA is un-accessible")
     public void verifyExceptionWhenJiraIsUnreachable() {
         Assertions.assertThrows(JiraClientRunTimeException.class,
-                () -> flowService.cxParseResults(basicScanRequest, getFileFromResourcePath()),
+                () -> sastScannerService.cxParseResults(basicScanRequest, getFileFromResourcePath()),
                 "Expected to get Jira un-accessible exception error");
     }
